replace tarfile-based functionality with simple file up/download
The problem is that if tarfile encounters Permission denied, it
ends the sending process, while the receiving side is still waiting
for more data to come in, unable to expect an error.

(Further, there's some python3.6 bug where tarfile sometimes adds
 extra zeros at the end, leaving them in the socket buffer, corrupting
 the data stream.)

The biggest issue here is that we have no "connections", we do not
start with a clean slate when executing/uploading/downloading,
we resume whatever was there before.

The second biggest issue is blocking - we can't even send out an
error from the server to the client, because if the client isn't
listening (crashed), writing the error packet would block, blocking
the server. If a new client tried to connect and (ie.) PING, then
they would receive the ERROR from the previous 'session'.

TL;DR - without vsock or some other support for connections,
it's very hard to do this cleanly in less than ~1000 lines of code.

Signed-off-by: Jiri Jaburek <comps@nomail.dom>