remove custom guest shell, use ssh again
While the final result worked well enough, the potential for easy
breakage caused by the lack of discrete 'connections', along with
extra workarounds that would be needed (for fapolicyd, which doesn't
allow execution of scripts that don't come from signed RPMs)
caused the custom shell to /not/ be a big improvement over regular ssh
with its own workarounds (rules we never remediate / fix afterwards).

And out of the two, ssh is probably more maintainable.

Signed-off-by: Jiri Jaburek <comps@nomail.dom>