add /lib and /conf to test requires
This is a new TMT feature, an optimization, where it soon stops
copying the entire repository tree to the remote system - only the
test directory is copied by default.

In order to have access to PYTHONPATH exported in individual tests,
we need to specify /lib (and /conf for waivers) as required.

The relative path locations should be preserved this way, according
to TMT developers.

Signed-off-by: Jiri Jaburek <comps@nomail.dom>