decouple Beaker and FMF wrapper, support pure restraint runs
This gets rid of the concept of "running in Beaker" or "running in TMT"
(mostly), replacing it with API-specific checks, decoupling results
reporting and test name retrieval.

The reason is that multiple infrastructure pieces use the FMF wrapper
(Beaker, Beaker-less restraint, etc.), so we can in fact end up with:

- Beaker results API (LAB_CONTROLLER), and FMF wrapper
- no Beaker results API, and FMF wrapper (standalone restraint)
- TMT results API, no FMF wrapper

So I modified the reporting code to just check if it can access Beaker
results API / TMT result API, and fall back to plain text.

Independently of this, test name is retrieved via TMT or FMF wrapper,
no matter whether it runs under Beaker or not.

Signed-off-by: Jiri Jaburek <comps@nomail.dom>