remove 'ssg' as a waiver reference to scap-security-guide
This is because we can no longer rely on a SSG RPM to be the source
of content, given that it can be provided via environment variables
as arbitrary version-less locations.

We've migrated to using 'released' / 'upstream' waiver files
for this reason too, and 'ssg' as a version is not being used
anywhere in waivers anymore.

Signed-off-by: Jiri Jaburek <comps@nomail.dom>