add repositories via RpmPack in lib.osbuild
This impacts installed repositories only, it has no effect on image
building. The reason for this is that osbuild doesn't support
file:// in gpgkey URLs, it requires either
  - an HTTP URL
  - the full body of the GPG key (ascii armored)

While we could use the second option, it would result in
the Blueprint being very large and impossible to debug,
so just add .repo files via RpmPack to the image.

Signed-off-by: Jiri Jaburek <comps@nomail.dom>