do not add custom partitions to osbuild Blueprint
These are already generated by 'oscap xccdf generate fix'.

The sizes used by 'oscap' are very big, but since the final image
is qcow2 with pre-allocation disabled (default for qcow2), we don't
really need to care as long as the image doesn't grow beyond
reasonable size (ie. from 800MB to 20+ GB).

Since we don't store tons of data in the Guest, this is a non-issue.

Signed-off-by: Jiri Jaburek <comps@nomail.dom>