remove TODO excludes in conf.remediation
These were added from a former Beakerlib test and have not proved
to be relevant to Contest, and likely never will.

Plus, this fixes a bug for ansible tests which had 'excludes'
overriden with an empty list (a '=' was used, not '+='), though
since no code was above 'Hardenings via Ansible', the bug never
manifested.

Signed-off-by: Jiri Jaburek <comps@nomail.dom>