globally unselect accounts_password_set_max_life_existing
Not only this lets us get rid of a lot of code, spread across several
places (the 'chage' should have been in tests, not in lib.virt),
it also fixes /hardening/image-builder which, so far, hasn't run
into the issue due to most profiles causing an error.

For image-builder, there wouldn't have been a chage-style fix
possible anyway, as even RpmPack scriptlets run before the
oscap-based hardening. There's no reasonable way to run anything
after the hardening, and before first boot.

It also rids us of the ever-changing syntax of
/etc/sysconfig/qemu-ga as the only qemu-g-a command remaining
is a soft reboot, which is allowed by default.

Signed-off-by: Jiri Jaburek <comps@nomail.dom>