fully remove image-builder GUI tests
https://issues.redhat.com/browse/RHEL-24246 is unlikely
to be fixed anytime soon, and it's not a compliance related
issue - osbuild doesn't support Server* package groups
in general.

The fact that the bug not being fixed has never come up for
image-builder (osbuild) users means GUI use cases are likely
non-existent with image-builder.

So let's not complicate our lives with testing something
seemingly so unusual.

Signed-off-by: Jiri Jaburek <comps@nomail.dom>