remove /static-checks/html-links from the daily/weekly plans
The test frequently fails on issues that are typically resolved
within a few days.

While a better solution would be able to detect long-term changes
to links rather than just closing our eyes for daily/weekly runs,
the exclude is a lot easier to implement and doesn't require
a dedicated process.

Signed-off-by: Jiri Jaburek <comps@nomail.dom>