move osbuild blueprint generation to the test
This adds translate_oscap_blueprint() and calls 'oscap' similarly to
/hardening/kickstart, showing the base test algorithm directly in
the test.

Coincidentally, this also cleans up the lib.osbuild logic a lot,
as we no longer have to pass 'profile' around and account for special
oscap cases and blueprint paths.
All oscap-specific stuff is now either in the translate function,
or directly in the test.

The TEMPLATE and other pieces of code were inspired by lib.virt.

translate_oscap_blueprint() weirdly takes lines instead of a full
block of text (the first thing it does is 'join()'), but that's
intentional, to keep the API similar to lib.virt translate functions.

Signed-off-by: Jiri Jaburek <comps@nomail.dom>