start using PyYAML instead of hacking YAML syntax
This was likely a RHEL-7 limitation, plus we didn't want to add
unnecessary dependency on non-python-stdlib modules (like 'requests'
or 'PyYAML').

However we keep running into the need to parse YAML (ie. playbooks
or TMT metadata), and python3*-pyyaml seems to be widely available
in RHEL (even ie. python39-pyyaml on RHEL-8), unlike 'rpm' or 'dnf',
so we might as well use it.

The '/static-checks/ansible/allowed-modules' test already uses it
without adding a RPM dependency, which hints at how widely available
PyYAML is. Also, TMT probably depends on it too.

Signed-off-by: Jiri Jaburek <comps@nomail.dom>