build CONTEST_CONTENT in a TMT plan prepare step
This is because building takes a long time (10+ minutes) on some
slow systems and that easily exceeds a 5/10 minute 'duration'
of a small test that never anticipated the extra delay.

Moving it to a plan moves it outside the 'duration' count.

Also, it's probably a better central place given that all tests
always share the built content so the ./build_product options
have to be considered globally.

Finally, this is not a great fix given that we are still left
with the 'build' argument to get_content(),

    def get_content(build=True):
        ...

and its SRPM-based building code, which still runs within test
'duration', however there is no easy solution for that (other than
getting rid of SRPM support entirely), so this commit at least
tries to bump things in a right direction.

Signed-off-by: Jiri Jaburek <comps@nomail.dom>