pass image/format as arguments to virt.import_image()
I originally rejected this because import_image() would be copying
the image file if it wasn't in /var/lib/libvirt/images, and it seemed
better and more explicit to make the user put the image on the right
place (into /var/lib/libvirt/images).

However, an alternate view point on "import" is "just use the disk
image in-place, wherever it is", which is what the function does now
with this change.

In that case, passing arguments makes more sense than setting instance-
wide variables.

Signed-off-by: Jiri Jaburek <comps@nomail.dom>