split virt.install() into install() and install_basic()
The idea is that install_basic() is a very minimal wrapper for
running Anaconda with the provided kickstart, no extra bells.

It is then install() that provides OS customizations for testing
via RpmPack, installs DNF repositories from the host, etc., etc.

I felt this was slightly better than trying to re-introduce the
verbatim_kickstart=False flag for install(), or something similar.
Splitting the functions also seems more readable, as we can now
use 'with' instead of an ExitStack.

The class Kickstart (packages, assembly) changes are also related
here, as this change is really about making install_basic() as minimal
as feasible, without forced %packages.

The intended use case is for 'ostreecontainer', where host repositories
and RpmPack don't make sense, given that the OS has already been
hardened and customized in the podman container.

Signed-off-by: Jiri Jaburek <comps@nomail.dom>