Improve waiver related to systemwide crypto policy
The rule configure_crypto_policy fails in test /hardening/anaconda/ospp
on RHEL 8 because the OSPP profile requires the policy to be set to
FIPS:OSPP, but the actual state of the policy is FIPS. This is a known
bug in Anaconda: https://issues.redhat.com/browse/RHEL-4722. For the
same reason the rule enable_fips_mode fails the same test as well.