remove 'openssh-server' from lib.virt requires
This breaks content remediations when GUEST_REQUIRES are added
to RpmPack and then a remediation does

  dnf -y remove openssh-server
  dnf -y install openssh-server

as the first command removes the RpmPack, removing also all system
dnf repositories.

I can't find why I added 'openssh-server' as a kickstart dependency
back when refactoring startup scripts, but I don't see a good
reason for it to be there - I guess the idea was (at some point)
to enforce our RpmPack RPM scripts to run later in the transaction,
but I again can't find anything that would rely on it anymore.

Signed-off-by: Jiri Jaburek <comps@nomail.dom>