Align FIPS environment settings in GUI test
Align FIPS environment settings so that the WITH_FIPS is set
to 1 in all GUI tests where it's set in the corresponding non-GUI
test (non-GUI tests are defined in main.fmf whereas GUI tests are
defined in with-gui.fmf). The exception is STIG where we have
a special stig_gui profile where the WITH_FIPS environment variable
is already set.