Account for SSG:DISA N:1 / 1:N mapping of rules in disa-alignment
The DISA Alignment test had problem that it doesn't account for
situations when a single SSG rule maps to multiple DISA rules
or when a single DISA rule maps to multiple SSG rules.

In these situations we get false positive results. This commit
attempts to fix these issues. The test will not use compare_results.py
from ComplianceAsCode because that script is too complex but we will
process the results by functions in scanning/disa-alignment/shared.py.

Fixes: ComplianceAsCode/content#13020

We will also remove the waivers of the upstream issue.