don't wait for ssh on virt snapshot restore
By definition, once restored, the sshd is reachable.

There is a corner case when the VM has been snapshotted (but offline)
for a longer period of time when the ARP table entry expires, but in
that case, 'ssh' will just take a few ms longer, its TCP connection
is designed to retry if necessary.

(The extra waiting logic doesn't help anything.)

Same for IP address - the restored guest doesn't try to get a new IP,
unbound (libvirtd dhcp server) already assigned it one, with a long
enough lease:

  NETWORK_EXPIRY = 168   # hours

Signed-off-by: Jiri Jaburek <comps@nomail.dom>