allow 'content_dir' override for util.* content functions
(Sorry to yet again increase the complexity here.)

There is unfortunately a valid use case for this: downloading SRPM for
an installed scap-security-guide RPM, extracting it, building it in
some special way, and wanting to access content inside that build dir.

This is not covered by CONTEST_CONTENT, or any other functionality.

The change in this commit should hopefully be minimally invasive,
while allowing the override.

I also un-private-d some functions since they can be useful even
outside the file, when trying to work with built content.

Signed-off-by: Jiri Jaburek <comps@nomail.dom>