speed up Guest.snapshotted(), avoid waiting for ipaddr
The waiting is needed only when the Guest instance doesn't have it yet,
ie. by reusing a snapshot from a different test.

If we already do have it, it won't change by the virtue of very long
DHCP lease times (as set in NETWORK_EXPIRY), so there's no point in
waiting for it every time.

Signed-off-by: Jiri Jaburek <comps@nomail.dom>