Resolve crypto policy rules alignment with DISA
This change will resolve https://github.com/ComplianceAsCode/content/issues/13108.

In the current upstream content, the only rule reported by
disa-alignment test as misaligned with DISA is configure_crypto_policy
on RHEL 9. But, the misalignment is reported because the testing VM
doesn't run in FIPS mode. (See the
https://github.com/ComplianceAsCode/content/issues/13108 for more
explanation.)

The STIG profiles require the system to be run in FIPS mode.
We think that to get more natural results in the disa-alignment test we
should install the VM on which the test runs in FIPS mode as well.

Running the testing VM in FIPS mode will cause that the misalignment
on configure_crypto_policy won't be reported by the test.