Differentiate remediation types
Update the `has_remediation()` function to differentiate between various
remediation types and the `lib/waive.py` module in a way that it is possible
to waive test results for rules which don't have any specified remediation
type.

Also update the `scripts/find_invalid_waivers.py` script to ignore and don't
mark waivers which use the `no_remediation` function for removal as we always
want to keep them.

Resolves #87