_rule_has_no_remediation: don't check error results
Error results might contain arbitrary strings in the test name
that might not be SCAP rules. This would make
`oscap.global_ds().has_remediation` to error out becase it tries
to extract a SCAP rule name from the test name and it also checks
if the extracted SCAP rule is present in the tested datastream.