/static-checks/rule-identifiers: use references from datastream
The references are defined inside a datastream file under 'Benchmark/reference'
so we don't need to store them statically in the test.

The mapping between profiles and references is not defined in datastream
nor in CaC/content build system so we keep it in the test code.

Also add checking for `ccn` and `bsi` references, and check for `ism_o`
profile on RHEL<=9 and for `ism_o_top_secret` on RHEL>=10.