Fix disa-alignment test
This test doesn't work and everything passes, because it fails
to find the STIG references in all rules. The root cause is the
recent change of the STIG URL from https://public.cyber.mil
to https://www.cyber.mil in our content. The old URL was hardcoded
in the test. To prevent problems like this in future, we will
parse the URL from the data stream instead of defining it in the
test.